refactor: SelectTag entity 기준으로 폴더 및 파일명 수정 => TagSelect
