+++ src/pages/PostsManagerPage.tsx
@@ -8,7 +8,7 @@ import { PostAddModal } from "../features/post-add/ui/PostAddModal"
 import { usePostsContext } from "../shared/model/PostContext"
 import { PostTable } from "../features/post-view/ui/PostTable"
 import { Pagination } from "../features/post-view/ui/Pagination"
-import { SelectTag } from "../features/post-filter/ui/SelectTag"
+import { SelectTag } from "../features/tag-select/ui/TagSelect"
 import { SelectSortStandard } from "../features/post-sort/ui/SelectSortStandard"
 import { SelectSortOrder } from "../features/post-sort/ui/SelectSortOrder"
 import { PostEditModal } from "../features/post-edit/ui/PostEditModal"
